[feat] Add policy-based authorization
Added policy-based authorization. Moved oauth logic to service.
