+++ DotNetflixServer/Services/AuthService/AuthServiceImpl.cs
@@ -68,6 +68,8 @@ public class AuthServiceImpl : IAuthService
             return new AuthResultDto(creatingResult.Errors.FirstOrDefault()?.Description);
         }
 
+        await _userManager.AddClaimAsync(user, new Claim("level", "user"));
+
         return new AuthResultDto();
     }
