--- DotNetflixServer/Services/GoogleOAuth/GoogleOAuth.cs
@@ -1,7 +1,0 @@
-﻿namespace Services.GoogleOAuth;
-
-public class GoogleOAuth
-{
-    public string ClientId { get; set; }
-    public string ClientSecret { get; set; }
-}
