--- DotNetflixServer/Services/OAuthService/IOAuthService.cs
@@ -1,6 +1,0 @@
-﻿namespace Services.OAuthService;
-
-public interface IOAuthService
-{
-    
-}
