--- DotNetflixServer/Services/OAuthService/OAuthServiceImpl.cs
@@ -1,6 +1,0 @@
-﻿namespace Services.OAuthService;
-
-public class OAuthServiceImpl : IOAuthService
-{
-    
-}
